Project to demonstrate TTD of fixed length 39 characters
This project is a simple COnsole application using MSTest to demonstrate
the value of using TTD to enhance the quality of an application.  AS we
can see in this solution, a developer can add as much tests as possible
allowing the design to evolve from simple tests. This helps the code to
evolve in quality as well as in semantic ways